requirements.txt
streamlit>=1.32.0
anthropic>=0.25.0
pandas>=2.0.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
